requirements.txt
aiohttp==3.11.14
openai==1.69.0
pandas==2.2.3
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
tenacity==9.0.0
tiktoken==0.9.0
torch==2.6.0
transformers==4.50.2
matplotlib~=3.9.4